Fixed bug in adopt_perspective
Changed the adopt_perspective and abduce plans. The abducible atoms have to be derived reaching the penultimate level of recursion. Before adopting the BB of the agent at the last level of recursion, the abducible atoms have to be derived, and then abducible/1 have to excluded (in order not to interfere with the abductive reasoning process).